Improvements
* list_all() on the confluent client and accepts multiple resources class
* more attributes for clusterv2, missing from the docs

+++ src/confluent_cloud_sdk/confluent_cluster_v2.py
@@ -3,9 +3,13 @@
 
 from __future__ import annotations
 
+import json
 import warnings
-from typing import Union
+from typing import List, Optional, Union
 
+from pydantic import BaseModel, Extra
+
+from .confluent_cloud_api import ResourcesList
 from .confluent_cloud_api.cluster_v2 import SpecModel as ClusterV2
 
 
@@ -15,18 +19,27 @@ class KafkaClusterV2:
     `CMKv2 <https://docs.confluent.io/cloud/current/api.html#tag/Clusters-(cmkv2)>_`
     """
 
-    api_v2_path = "/cmk/v2"
-    clusters_path = f"{api_v2_path}/clusters"
+    api_path = "/cmk/v2/clusters"
 
     def __init__(
         self,
         client,
-        environment_id: str,
+        resource_id: str = None,
+        environment_id: str = None,
+        spec: dict = None,
     ):
-        self.api_path = self.clusters_path
         self._client = client
         self._resource = None
         self._resource_class = ClusterV2
+
+        if resource_id and environment_id and not spec:
+            req = client.get(
+                f"{self._client.api_url}{self.api_path}/{resource_id}?environment={environment_id}"
+            )
+            # print(json.dumps(req.json(), indent=2))
+            self._resource = self._resource_class(**req.json())
+        elif spec:
+            self._resource = self._resource_class(**spec)
 
     @property
     def resource(self) -> ClusterV2:
@@ -38,18 +51,12 @@ class KafkaClusterV2:
             return self._resource.id.__root__
         return None
 
-    def list(self, environment_id: str) -> list:
-        if self.api_path:
-            return self._client.get(
-                f"{self._client.api_url}{self.api_path}?environment={environment_id}"
-            )
-        return []
-
     def read(self):
         return self._client.get(self.resource)
 
     def update(self, description: str):
         warnings.warn(NotImplemented("Update cluster not implemented"))
 
-    def delete(self):
-        return self._client.delete(self.resource.metadata.self)
+    def delete(self, no_dry_run: bool = False):
+        if no_dry_run:
+            return self._client.delete(self.resource.metadata.self)
